requirements.txt
Django==2.1.7
django-cors-middleware==1.3.1
django-extensions==2.1.5
djangorestframework==3.9.1
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
